backend/requirement.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
Pillow==10.2.0
pdf2image==1.16.3
email-validator